% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{aggregateCountries}
\alias{aggregateCountries}
\title{aggregateCountries}
\usage{
aggregateCountries(con, df, groups)
}
\arguments{
\item{con}{(DBI)}

\item{df}{(dataframe) events or population}

\item{groups}{(dataframe) a country-group mapping dataframe}
}
\description{
aggregate countries by group
}
