% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_data.R
\name{assemble_data_plot_predictions}
\alias{assemble_data_plot_predictions}
\title{assemble_data_plot_predictions}
\usage{
assemble_data_plot_predictions(
  predictions,
  ts_data,
  days_predict,
  model,
  training_window,
  confidence
)
}
\arguments{
\item{predictions}{(list): the output of get_predictions()}

\item{ts_data}{(dataframe): an events dataframe}

\item{days_predict}{(numeric): How many days in the future to predict}

\item{model}{(list): the output of get_best_model()}

\item{confidence}{(numeric): Which level of confidence we want to keep. 
One of 80, 95 or NULL. If NULL, we do not keep any confidence}

\item{train_window}{(numeric): the number of days prior to the last}
}
\value{
a list with the predictions and the confidence intervals at 80% and 95%
}
\description{
Clean up and format the data
}
