% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{assembleType}
\alias{assembleType}
\title{assembleType}
\usage{
assembleType(type)
}
\arguments{
\item{type}{(character) One of "cases", "deaths" or "recovered"}
}
\value{
data frame
}
\description{
Create a dataframe for one of the COVID variables: cases, deaths or recovered
}
