% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesDef.R
\docType{class}
\name{covidData-class}
\alias{covidData-class}
\title{Data}
\description{
Data
}
\section{Slots}{

\describe{
\item{\code{JHUData_raw}}{data.frame of settings}

\item{\code{JHUData_smooth}}{data.frame}

\item{\code{JHUData_diffSmooth}}{data.frame}

\item{\code{demographic}}{sf}
}}

