% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{diffRawDf}
\alias{diffRawDf}
\title{diffRawDf}
\usage{
diffRawDf(df)
}
\arguments{
\item{df}{(data.frame) a data frame with rows of time series}
}
\value{
diff'ed data frame
}
\description{
diff a df with time series
}
