% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{diffSmoothDf}
\alias{diffSmoothDf}
\title{diffSmoothDf}
\usage{
diffSmoothDf(df)
}
\arguments{
\item{df}{(data.frame) a data frame with rows of time series}
}
\value{
diff'ed data frame
}
\description{
diff a df with time series
}
