% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{displayEvents}
\alias{displayEvents}
\title{displayEvents}
\usage{
displayEvents(events)
}
\arguments{
\item{events}{(dataframe)}
}
\value{
dataframe
}
\description{
format events table to be displayed in the app
}
