% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{displayPopulation}
\alias{displayPopulation}
\title{displayPopulation}
\usage{
displayPopulation(population, cols)
}
\arguments{
\item{cols}{(character) variables to be included. If NULL,
all variables will be included}

\item{events}{(dataframe)}
}
\value{
dataframe
}
\description{
format population table to be displayed in the app
}
