% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMaps.R
\name{doMap}
\alias{doMap}
\title{doMap}
\usage{
doMap(df, doFacet)
}
\arguments{
\item{df}{(data.frame) the dataframe from getEventsMapDb()}

\item{doFacet}{(logical) TRUE if faceting the plot}
}
\value{
tmap object
}
\description{
plot in amp chart
}
