% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDay_normalised.R
\name{doMapData_normalised}
\alias{doMapData_normalised}
\title{doMapData_normalised}
\usage{
doMapData_normalised(
  world,
  filterByCountry = NULL,
  plotMetric = "cases",
  chosenDay = NULL
)
}
\arguments{
\item{world}{(sf): sf object to plot}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{plotMetric}{(character) One of "cases", "deaths", "recovered"}

\item{chosenDay}{(numeric): the day to plot}
}
\value{
a ggplot
}
\description{
plot on a map
}
