% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGBQuarantine_binary.R
\name{doMapGBQuarantine_binary}
\alias{doMapGBQuarantine_binary}
\title{doMapGBQuarantine_binary}
\usage{
doMapGBQuarantine_binary(
  plotData,
  filterByCountry = NULL,
  plotMetric = "cases",
  chosenDay = NULL
)
}
\arguments{
\item{plotData}{(covidData): S4 object}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{plotMetric}{(character) One of "cases", "deaths", "recovered"}

\item{chosenDay}{(numeric): the day to plot}
}
\value{
a ggplot
}
\description{
plot on a map
}
