% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{doPlot}
\alias{doPlot}
\title{doPlot}
\usage{
doPlot(df, typePlot, country = NULL)
}
\arguments{
\item{typePlot}{cases, deaths or recovered}

\item{country}{the country to plot}

\item{data}{df of data from JHU}
}
\description{
plot data for a single country
}
