% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsFun.R
\name{eventsPlot}
\alias{eventsPlot}
\title{eventsPlot}
\usage{
eventsPlot(events, population, normBy = NULL, multiplyFactor = 1)
}
\arguments{
\item{events}{(data.frame) events dataframe}

\item{population}{(data.frame) population dataframe}

\item{normBy}{(character) name of the column to use in population}

\item{multiplyFactor}{(numeric) positive integer}
}
\value{
ggplot
}
\description{
plot events data
}
