% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{fillNAs}
\alias{fillNAs}
\title{fillNAs}
\usage{
fillNAs(df)
}
\arguments{
\item{df}{(data.frame) dataframe}
}
\value{
population data frame
}
\description{
replace NAs in numerical columns. Used with the events data.frame. NAS
are replacing using zoo::na.approx. Leading NAs are replaced
with 0s, trailing NAs are replaced with the last non-NA value.
}
