% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_data.R
\name{get_best_model}
\alias{get_best_model}
\title{get_best_model}
\usage{
get_best_model(ts_data, train_window)
}
\arguments{
\item{ts_data}{(dataframe): an events dataframe}

\item{train_window}{(numeric): the number of days prior to the last 
observation to use for training}
}
\value{
a list with model_name, model_fit and standard_error
base', # base is currently disabled
}
\description{
Find the best model between the implemented one: 'base',
'snaive', 'ets', arima'. The best model is the one with the lowest residuals
}
