% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_data.R
\name{get_predictions}
\alias{get_predictions}
\title{get_predictions}
\usage{
get_predictions(model, days_predict)
}
\arguments{
\item{model}{(list): the output of get_best_model()}

\item{days_predict}{(numeric): How many days in the future to predict}
}
\value{
a list with the predictions and the confidence intervals at 80% and 95%
}
\description{
Calculate the predictions from the model
}
