% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_data.R
\name{get_residuals}
\alias{get_residuals}
\title{get_residuals}
\usage{
get_residuals(model_name, ts_data)
}
\arguments{
\item{model_name}{(character): The name of the model to use. One of 'base',
'snaive', 'ets', arima'. NOTE: 'base' currently disabled due to poor of accuracy}

\item{ts_data}{(dataframe): an events dataframe}
}
\value{
a list with model_name, model_fit and standard_error
}
\description{
Get the residuals of a model. We use this metric to decide the
best model (lower residual is better). Future implementations might include AIC
}
