% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getAllData}
\alias{getAllData}
\title{getAllData}
\usage{
getAllData()
}
\value{
data frame
}
\description{
download data from 'https://covid.ourworldindata.org/data/owid-covid-data.xlsx'
and import it as a dataframe
}
