% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getAreasForMap}
\alias{getAreasForMap}
\title{getgetAreasForMapPopulation}
\usage{
getAreasForMap(possibleCountries)
}
\arguments{
\item{possibleCountries}{(character) vector of getAllData()$Country}
}
\value{
character vector of countries only
}
\description{
remove aggregated areas from the map as we plot
only at country level
}
