% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMaps.R
\name{getAvailableCountries}
\alias{getAvailableCountries}
\title{getAvailableCountries}
\usage{
getAvailableCountries(con)
}
\arguments{
\item{con}{(DBI)}
}
\value{
dataframe
}
\description{
getAvailableCountries
}
