% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{getChosenDate}
\alias{getChosenDate}
\title{getChosenDate}
\usage{
getChosenDate(df, idxchosenDay)
}
\arguments{
\item{idxChosenDay}{(numeric): index of he column with the name to covert to character date}
}
\value{
a character date
}
\description{
return character form of the day to plot given as index
}
