% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getCountriesFromGroups}
\alias{getCountriesFromGroups}
\title{getCountriesFromGroups}
\usage{
getCountriesFromGroups(groups)
}
\arguments{
\item{groups}{(character) vector of groups to retrive}
}
\value{
data frame
}
\description{
map groups to countries
}
