% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getCovidDf}
\alias{getCovidDf}
\title{getCovidDf}
\usage{
getCovidDf(df)
}
\arguments{
\item{df}{(data.frame) the data frame to convert to a list}
}
\value{
list with keys, populationDf and dataDf
}
\description{
create a covidDf object
}
