% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(rootData, fileName, typeIs)
}
\arguments{
\item{rootData}{(character) the raw data from JHU}

\item{fileName}{(character) the name of the file with the raw data}

\item{typeIs}{(character) one of "cases", "deaths", "recovered"}
}
\value{
data frame of cases, deaths and recovered
}
\description{
prep data by reading in thhe downloaded repo,
joining coubtries with more than one area and aggregating
}
