% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getEvents}
\alias{getEvents}
\title{getEvents}
\usage{
getEvents(allData)
}
\arguments{
\item{allData}{(dataframe) obtained from getAllData()}
}
\value{
data frame
}
\description{
prepare the events table for the database
}
