% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMaps.R
\name{getEventsMapDb}
\alias{getEventsMapDb}
\title{getEventsMapDb}
\usage{
getEventsMapDb(countries, metric, date, normBy, multiplyFactor)
}
\arguments{
\item{countries}{(character) vector of countries to plot}

\item{metric}{(character) the metric to plot}

\item{date}{(character) the date to plot in
the format "YYYY-MM-DD"}

\item{normBy}{(character) name of the column to use in population}

\item{multiplyFactor}{(numeric) positive integer}
}
\value{
normalised events data.frame
}
\description{
get the data for the map
}
