% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getGlobal}
\alias{getGlobal}
\title{getGlobal}
\usage{
getGlobal(pathGlobal)
}
\arguments{
\item{pathGlobal}{(character) the path to where the global data is}
}
\value{
data frame
}
\description{
get the global data
}
