% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getGroups}
\alias{getGroups}
\title{getGroups}
\usage{
getGroups(df)
}
\arguments{
\item{df}{(data.frame) from getAllData()}
}
\value{
data frame
}
\description{
get groups of countries, i.e. europe vs oceania etc
}
