% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{getIdxChosenDay}
\alias{getIdxChosenDay}
\title{getIdxChosenDay}
\usage{
getIdxChosenDay(df, chosenDay)
}
\arguments{
\item{df}{(dataframe) JHU dataframe}

\item{chosenDay}{(numeric): the day to plot}
}
\value{
numeric index
}
\description{
return the dataframe index for the chosen day
}
