% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getKeys}
\alias{getKeys}
\title{getKeys}
\usage{
getKeys(populationDf)
}
\arguments{
\item{populationDf}{(data.frame) the populationDf of the covidData object}
}
\value{
character vector
}
\description{
Get all the possible keys for country/state/city
}
