% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMaps.R
\name{getMapData}
\alias{getMapData}
\title{getMapData}
\usage{
getMapData(con, events)
}
\arguments{
\item{con}{(DBI)}

\item{events}{(dataframe)}
}
\value{
dataframe
}
\description{
get the data for the map plot
}
