% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGBQuarantine_binary.R
\name{getMapGBQuarantine_binary}
\alias{getMapGBQuarantine_binary}
\title{getMapGBQuarantine_binary}
\usage{
getMapGBQuarantine_binary(
  plotData,
  filterByCountry = NULL,
  plotMetric = "cases",
  chosenDay = NULL
)
}
\arguments{
\item{plotData}{(covidData): S4 object}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{plotMetric}{(character) One of "cases", "deaths", "recovered"}

\item{chosenDay}{(numeric): the day to plot}
}
\value{
data frame of cases, deaths and recovered
}
\description{
get the data to plot on a map of total number of cases
}
