% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{getMapTitle}
\alias{getMapTitle}
\title{getMapTitle}
\usage{
getMapTitle(
  world,
  filterByCountry = NULL,
  plotMetric = "cases",
  chosenDay = NULL,
  plotType
)
}
\arguments{
\item{world}{(sf): results of getMap[...]()}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{plotMetric}{(character) One of "cases", "deaths", "recovered"}

\item{chosenDay}{(numeric): the day to plot}

\item{plotType}{(character): one of doMapTrend_normalise, doMapTrend,
doMapDataRate_raw, doMapDataRate_normalised,
doMapGBQuarantine_binary, doMapGBQuarantine, doMapData_raw,
doMapData_normalised}
}
\value{
character string of the map title
}
\description{
Get the map title. Used by the frontend app
}
