% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.R
\name{getNewGADM}
\alias{getNewGADM}
\title{getNewGADM}
\usage{
getNewGADM(dataObj)
}
\arguments{
\item{dataObj}{(coviData)}
}
\value{
save country rds to the data folder
}
\description{
refresh GADM map data for all needed couontries
}
