% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getOptions}
\alias{getOptions}
\title{getOptions}
\usage{
getOptions(con, table, field)
}
\arguments{
\item{con}{(DBI)}

\item{table}{(character) the name of the database table}

\item{field}{(character) the name of the field in the table}
}
\value{
a character vector
}
\description{
get all the unique values in the specified
field of the specified table
}
