% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getPopulation}
\alias{getPopulation}
\title{getPopulation}
\usage{
getPopulation(allData)
}
\arguments{
\item{allData}{(dataframe) obtained from getAllData()}
}
\value{
data frame
}
\description{
prepare the population table for the database
}
