% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsFun.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{getSummaryTable}
\usage{
getSummaryTable(df)
}
\arguments{
\item{df}{JHU data.frame}
}
\value{
a summary data.frame
}
\description{
summary table for all data obtained from JHU
}
