% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{normaliseEvents}
\alias{normaliseEvents}
\title{normaliseEvents}
\usage{
normaliseEvents(events, population, normBy, multiplyFactor = 1)
}
\arguments{
\item{events}{(data.frame) events dataframe}

\item{population}{(data.frame) population dataframe}

\item{normBy}{(character) name of the column to use in population}

\item{multiplyFactor}{(numeric) positive integer}
}
\value{
normalised events data.frame
}
\description{
divide the value in the events by the specified
normBy and multiply by a factor. For instance to normalise the data
by every 100 inhabitants, set normBy='population' and
multiplyFactor=100
}
