% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{normalizeByPop}
\alias{normalizeByPop}
\title{normalizeByPop}
\usage{
normalizeByPop(df)
}
\arguments{
\item{idxChosenDay}{(numeric): index of he column with the name to covert to character date}
}
\value{
a dataframe
}
\description{
normalize data by populations
}
