% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{plotAllMetrics}
\alias{plotAllMetrics}
\title{plotAllMetrics}
\usage{
plotAllMetrics(allDf, countryPlot, scale = "linear", align = FALSE)
}
\arguments{
\item{allDf}{list of all Df metrics}

\item{countryPlot}{character of a country}

\item{scale}{character for plot in linear or log scale}

\item{align}{logical If we should align by date of min number of cases/deaths/recovered
or the rate of change (true)}
}
\description{
plot cases, deaths and recovered in a single plot
}
