% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{plotMap}
\alias{plotMap}
\title{plotMap}
\usage{
plotMap(
  plotData,
  filterByCountry = NULL,
  plotMetric = "cases",
  chosenDay = NULL,
  plotType
)
}
\arguments{
\item{plotData}{(covidData): S4 object}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{plotMetric}{(character) One of "cases", "deaths", "recovered"}

\item{chosenDay}{(numeric): the day to plot}

\item{plotType}{(character): one of doMapTrend_normalise, doMapTrend,
doMapDataRate_raw, doMapDataRate_normalised,
doMapGBQuarantine_binary, doMapGBQuarantine, doMapData_raw,
doMapData_normalised}
}
\value{
data frame of cases, deaths and recovered
}
\description{
single entry point to plot map
}
