% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{plotQMap}
\alias{plotQMap}
\title{plotQMap}
\usage{
plotQMap(world, categoricalPlot)
}
\arguments{
\item{world}{(data.frame): The countries with population data, from the package rnaturalearthdata
countries are included}

\item{categoricalPlot}{(logical): colour the countries in red if they had more than
20 cases, blue otherwise}
}
\value{
a ggplot
}
\description{
plot on a map the increase in number of cases over 100k subjects
over the past 7 days
}
