% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsFun.R
\name{populationForDisplay}
\alias{populationForDisplay}
\title{populationForDisplay}
\usage{
populationForDisplay(population)
}
\arguments{
\item{population}{(data.frame) population dataframe}
}
\value{
population data frame
}
\description{
remove columns of all nas
}
