% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsFun.R
\name{populationPlot}
\alias{populationPlot}
\title{populationPlot}
\usage{
populationPlot(population, variables)
}
\arguments{
\item{population}{(data.frame) population dataframe}

\item{variables}{(character) vector of column names to plot}
}
\value{
ggplot
}
\description{
plot population data
}
