% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_data.R
\name{predict_data}
\alias{predict_data}
\title{assemble_data_plot_predictions}
\usage{
predict_data(
  events,
  days_predict = 60,
  training_window = 14,
  confidence = NULL
)
}
\arguments{
\item{events}{(dataframe): an events dataframe}

\item{days_predict}{(numeric): How many days in the future to predict}

\item{confidence}{(numeric): Which level of confidence we want to keep. 
One of 80, 95 or NULL. If NULL, we do not keep any confidence}

\item{train_window}{(numeric): the number of days prior to the last}
}
\value{
a list with the predictions and the confidence intervals at 80% and 95%
}
\description{
Clean up and format the data
}
