% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{refreshData}
\alias{refreshData}
\title{refreshData}
\usage{
refreshData(getNew = TRUE, doSmoothing = TRUE)
}
\value{
data frame of cases, deaths and recovered
}
\description{
refresh data from https://github.com/CSSEGISandData/COVID-19.git
}
