% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{removeGroups}
\alias{removeGroups}
\title{removeGroups}
\usage{
removeGroups(allData, groups)
}
\arguments{
\item{allData}{(dataframe) from getAllData()}

\item{groups}{(dataframe) from}
}
\description{
remove entries like asia, africa,
europe etc from the list of countries
}
