% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{saveAllData}
\alias{saveAllData}
\title{saveAllData}
\usage{
saveAllData()
}
\value{
nothing
}
\description{
save a fresh allData covid object as RData
This is used in the app to refresh the stored data used to speed
up app loading
}
