% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{showGeographyFilter}
\alias{showGeographyFilter}
\title{showGeographyFilter}
\usage{
showGeographyFilter(dataObj)
}
\arguments{
\item{plotData}{(covidData): S4 object}
}
\value{
character vector
}
\description{
return the avilable geography filters
}
