% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{smoothDf}
\alias{smoothDf}
\title{smoothDf}
\usage{
smoothDf(df)
}
\arguments{
\item{df}{(data.frame) a data frame woth rows of time series}
}
\value{
smoothed data frame
}
\description{
smooth a df with time series
}
