% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{smoothValues}
\alias{smoothValues}
\title{smoothValues}
\usage{
smoothValues(myVal)
}
\arguments{
\item{myVal}{(numeric vectro): the numbers to smooth}
}
\value{
A numeric vector
}
\description{
smooth time series
}
