% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{sumCountries}
\alias{sumCountries}
\title{sumCountriesDf}
\usage{
sumCountries(df, populationDf)
}
\arguments{
\item{df}{(dataframe) the JHU dataframe}

\item{populationDf}{(dataframe) the population dataframe from the covidData object}
}
\value{
dataframe
}
\description{
summarise the rows of a JHU dataframe by country.
}
