% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{sumPast7}
\alias{sumPast7}
\title{sumPast7}
\usage{
sumPast7(df, idxchosenDay)
}
\arguments{
\item{df}{(dataframe)}

\item{idxchosenDay}{(numeric)}
}
\value{
a dataframe
}
\description{
normalize data by populations
}
