% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{updateDb}
\alias{updateDb}
\title{updateDb}
\usage{
updateDb(con)
}
\arguments{
\item{con}{(DBI)}
}
\value{
nothing
}
\description{
fetch new data from 'https://covid.ourworldindata.org/data/owid-covid-data.xlsx'
and update the database
}
