% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{doMap}
\alias{doMap}
\title{doData}
\usage{
doMap(
  plotData,
  normalizeByPopulation = FALSE,
  filterByCountry = NULL,
  removeCountries = NULL,
  chosenDay = NULL,
  showTrend = FALSE
)
}
\arguments{
\item{plotData}{(covidData): S4 object}

\item{normalizeByPopulation}{(logical): if TRUE, divide the number of cases by the
country population}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{removeCountries}{(character): countries to remove}

\item{chosenDay}{(numeric): the day to plot}

\item{showTrend}{(logical): colour the map by the variation of the last 5 days vs the previous
5 days}
}
\value{
a ggplot
}
\description{
plot on a map
}
