% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{doMap}
\alias{doMap}
\title{doData}
\usage{
doMap(
  rawData,
  normalizeByPopulation = FALSE,
  filterByCountry = NULL,
  chosenDay = NULL
)
}
\arguments{
\item{rawData}{(dataframe): the data obtaind from refresData()}

\item{normalizeByPopulation}{(logical): if TRUE, divide the number of cases by the
country population}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{chosenDay}{(numeric): the day to plot}
}
\value{
a ggplot
}
\description{
plot on a map
}
