% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{doMap}
\alias{doMap}
\title{doData}
\usage{
doMap(
  plotData,
  plotType,
  filterByCountry = NULL,
  plotMetric = "cases",
  chosenDay = NULL
)
}
\arguments{
\item{plotData}{(covidData): S4 object}

\item{plotType}{(character): one of avgPast7_raw, avgPast7_normalised,
cumulativeAll_raw, cumulativeAll_normalised, cumulativePast7_raw,
cumulativePast7_normalised, avgToday_raw, avgToday_normalised}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{plotMetric}{(character) One of "cases", "deaths", "recovered"}

\item{chosenDay}{(numeric): the day to plot}
}
\value{
a ggplot
}
\description{
plot on a map
}
