% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{doPlot}
\alias{doPlot}
\title{doPlot}
\usage{
doPlot(df, typePlot, countryPlot = NULL, scale = "linear", plotDiff = FALSE)
}
\arguments{
\item{typePlot}{cases, deaths or recovered}

\item{countryPlot}{the country to plot}

\item{scale}{if to plot in linear or log scale}

\item{plotDiff}{logical if we should plot the raw number (false)
or the rate of change (true)}

\item{data}{df of data from JHU}
}
\description{
plot data for a single country
}
