% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{doPlot}
\alias{doPlot}
\title{doPlot}
\usage{
doPlot(
  df,
  typePlot,
  countryPlot = NULL,
  scale = "linear",
  plotLim = NULL,
  align = FALSE
)
}
\arguments{
\item{df}{df of data from JHU}

\item{typePlot}{cases, deaths or recovered}

\item{countryPlot}{the country to plot}

\item{scale}{if to plot in linear or log scale}

\item{plotLim}{Dates max an min limits for the plot}

\item{align}{logical If we should align by date of min number of cases/deaths/recovered
or the rate of change (true)}
}
\description{
plot data for a single country
}
