% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsFun.R
\name{doPlot}
\alias{doPlot}
\title{doPlot}
\usage{
doPlot(df, reScale = FALSE)
}
\arguments{
\item{df}{(data.frame) events dataframe from getPlotData}

\item{reScale}{(boolean) if rescaling in the 0-1 range}
}
\value{
ggplot
}
\description{
plot events data
}
