% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(rootData, fileName, typeIs)
}
\value{
data frame of cases, deaths and recovered
}
\description{
prep data by reading in thhe downloaded repo,
joining coubtries with mopre than one area and aggregating
}
