% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getEventsDb}
\alias{getEventsDb}
\title{getEventsDb}
\usage{
getEventsDb(countries, metrics)
}
\arguments{
\item{countries}{(character) vector of countries to retrive}

\item{metrics}{(character) vector of metrics to retrive}
}
\value{
events data frame
}
\description{
retrive events from the database
}
