% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{getMapData}
\alias{getMapData}
\title{getMapData}
\usage{
getMapData(
  plotData,
  gadm,
  normalizeByPopulation = FALSE,
  filterByCountry = NULL,
  plotMetric = "cases",
  chosenDay = NULL,
  showTrend = FALSE
)
}
\arguments{
\item{plotData}{(covidData): S4 object}

\item{gadm}{(list) list with data of all the available countries}

\item{normalizeByPopulation}{(logical): if TRUE, divide the number of cases by the
country population}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{chosenDay}{(numeric): the day to plot}

\item{showTrend}{(logical): colour the map by the variation of the last 5 days vs the previous
5 days}

\item{plotMatric}{(character) One of "cases", "deaths", "recovered"}
}
\value{
data frame of cases, deaths and recovered
}
\description{
get the data to plot on a map
}
