% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getMapData}
\alias{getMapData}
\title{getMapData}
\usage{
getMapData(
  rawData,
  normalizeByPopulation = FALSE,
  filterByCountry = NULL,
  chosenDay = NULL
)
}
\arguments{
\item{rawData}{(dataframe): the data obtaind from refresData()}

\item{normalizeByPopulation}{(logical): if TRUE, divide the number of cases by the
country population}

\item{filterByCountry}{(character): The countries to plot in the map. If NULL all
countries are included}

\item{chosenDay}{(numeric): the day to plot}
}
\value{
data frame of cases, deaths and recovered
}
\description{
get the data to plot on a map
}
