% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlots.R
\name{getMapTitle}
\alias{getMapTitle}
\title{getMapTitle}
\usage{
getMapTitle(worldMap, plotMetric = "cases", chosenDay = NULL, plotType)
}
\arguments{
\item{plotMetric}{(character) One of "cases", "deaths", "recovered"}

\item{chosenDay}{(numeric): the day to plot}

\item{plotType}{(character): one of doMapTrend_normalise, doMapTrend,
doMapDataRate_raw, doMapDataRate_normalised,
doMapGBQuarantine_binary, doMapGBQuarantine, doMapData_raw,
doMapData_normalised}

\item{world}{(worldMap): S4 object of class worldMap}
}
\value{
character string of the map title
}
\description{
Get the map title. Used by the frontend app
}
