% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getPopulationDb}
\alias{getPopulationDb}
\title{getPopulationDb}
\usage{
getPopulationDb(con, groups, countries)
}
\arguments{
\item{con}{(DBI) database connection}

\item{groups}{(character) a geographical group of countries}

\item{countries}{(character) vector of countries to retrive}
}
\value{
population data frame
}
\description{
retrive population data from the database
}
