% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{plotAllMetrics}
\alias{plotAllMetrics}
\title{plotAllMetrics}
\usage{
plotAllMetrics(allDf, country)
}
\arguments{
\item{allDf}{list of all Df metrics}

\item{country}{character of a country}
}
\description{
plot cases, deaths and recovered in a single plot
}
