% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{plotAllMetrics}
\alias{plotAllMetrics}
\title{plotAllMetrics}
\usage{
plotAllMetrics(allDf, countryPlot, scale = "linear")
}
\arguments{
\item{allDf}{list of all Df metrics}

\item{countryPlot}{character of a country}

\item{scale}{character for plot in linear or log scale}
}
\description{
plot cases, deaths and recovered in a single plot
}
